Figuring out forwardRef and useRef
In the SignIn form, I want the username field to be focused on load. Using `forwardRef` to define the `HorizontalFieldInput` component lets me define a ref and pass it through to the actual input component (styled components is good about this; they're probably using forwardRef as well).

In the SignIn form, I can define the ref with `useRef` and then in `useEffect` I can focus it.

However, as I only want to focus it the first and only time, adding a state var for this seems pretty much overkill.